Rename generateReorderPointsMSD to generateReorderPoints as it will work for all POS now
